afterFirstReviewChanges

+++ Content.Client/SS220/SuperMatter/Emitter/SuperMatterEmitterBoltSystem.cs
@@ -27,6 +27,6 @@ public sealed class SuperMatterEmitterSystem : EntitySystem
             return;
 
         spriteComponent.Color = Colormaps.SMEmitter.GetCorrespondingColor(superMatterEmitter.EnergyToMatterRatio / 100f);
-        spriteComponent.Scale = new Vector2(MathF.Sqrt(superMatterEmitter.PowerConsumption / SuperMatterEmitterExtensionConsts.BaseEnergyConsumption));
+        spriteComponent.Scale = new Vector2(MathF.Sqrt(superMatterEmitter.PowerConsumption / (float)SuperMatterEmitterExtensionConsts.BaseEnergyConsumption));
     }
 }
